requirements.txt
dhanhq==2.0.2
python-dotenv==1.0.1
pandas==2.2.3
schedule==1.2.1
flask==3.0.3
sqlalchemy==2.0.31
psycopg2-binary==2.9.9
gunicorn==22.0.0
requests==2.32.3
scipy==1.13.1
